requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
langchain-community
huggingface-hub
InstructorEmbedding
sentence-transformers
fastapi
uvicorn
python-multipart
python-docx
nltk
requests
mangum
pymongo
pydantic
docling